# perf:reduce allocations in EncodingLoader