@HavenDV 2024-11-09
Released as 2.1.0 on NuGet